Adapt the schema registry to support built-in services
